Change package names
Rename mignonnesaurus to github.eliflores.
