+++ pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.mignonnesaurus.tools</groupId>
-    <artifactId>mignonnesaurus-little-tools</artifactId>
+    <groupId>com.github.eliflores.tools</groupId>
+    <artifactId>eliflores-little-tools</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
