+++ src/main/java/com/github/eliflores/tools/clients/CsvClientTool.java
@@ -1,0 +1,7 @@
+package com.github.eliflores.tools.clients;
+
+public class CsvClientTool {
+    public static void main(String[] args) {
+
+    }
+}
