+++ src/main/java/com/github/eliflores/tools/clients/XmlClientTool.java
@@ -1,0 +1,7 @@
+package com.github.eliflores.tools.clients;
+
+public class XmlClientTool {
+    public static void main(String[] args) {
+
+    }
+}
