+++ src/main/java/com/github/eliflores/tools/csv/CsvReader.java
@@ -1,0 +1,4 @@
+package com.github.eliflores.tools.csv;
+
+public class CsvReader {
+}
