+++ src/main/java/com/github/eliflores/tools/csv/CsvWriter.java
@@ -1,0 +1,4 @@
+package com.github.eliflores.tools.csv;
+
+public class CsvWriter {
+}
