+++ src/main/java/com/github/eliflores/tools/exception/MyLittleToolsException.java
@@ -1,0 +1,7 @@
+package com.github.eliflores.tools.exception;
+
+public class MyLittleToolsException extends RuntimeException {
+    public MyLittleToolsException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
