+++ src/main/java/com/github/eliflores/tools/json/model/Person.java
@@ -1,0 +1,22 @@
+package com.github.eliflores.tools.json.model;
+
+public class Person {
+    private String firstName;
+    private String lastName;
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+}
