+++ src/main/java/com/github/eliflores/tools/xml/XmlReader.java
@@ -1,0 +1,4 @@
+package com.github.eliflores.tools.xml;
+
+public class XmlReader {
+}
