--- src/main/java/com/mignonnesaurus/tools/clients/CsvClientTool.java
@@ -1,7 +1,0 @@
-package com.mignonnesaurus.tools.clients;
-
-public class CsvClientTool {
-    public static void main(String[] args) {
-
-    }
-}
