--- src/main/java/com/mignonnesaurus/tools/clients/XmlClientTool.java
@@ -1,7 +1,0 @@
-package com.mignonnesaurus.tools.clients;
-
-public class XmlClientTool {
-    public static void main(String[] args) {
-
-    }
-}
