--- src/main/java/com/mignonnesaurus/tools/csv/CsvReader.java
@@ -1,4 +1,0 @@
-package com.mignonnesaurus.tools.csv;
-
-public class CsvReader {
-}
