--- src/main/java/com/mignonnesaurus/tools/csv/CsvWriter.java
@@ -1,4 +1,0 @@
-package com.mignonnesaurus.tools.csv;
-
-public class CsvWriter {
-}
