--- src/main/java/com/mignonnesaurus/tools/exception/ExceptionUtil.java
@@ -1,10 +1,0 @@
-package com.mignonnesaurus.tools.exception;
-
-public final class ExceptionUtil {
-    private ExceptionUtil() {
-    }
-
-    public static MyLittleToolsException createException(String message, Throwable cause) {
-        return new MyLittleToolsException(message, cause);
-    }
-}
