--- src/main/java/com/mignonnesaurus/tools/exception/MyLittleToolsException.java
@@ -1,7 +1,0 @@
-package com.mignonnesaurus.tools.exception;
-
-public class MyLittleToolsException extends RuntimeException {
-    public MyLittleToolsException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
