--- src/main/java/com/mignonnesaurus/tools/url/UrlUtil.java
@@ -1,28 +1,0 @@
-package com.mignonnesaurus.tools.url;
-
-import com.mignonnesaurus.tools.exception.ExceptionUtil;
-
-import java.io.UnsupportedEncodingException;
-import java.net.URLDecoder;
-import java.net.URLEncoder;
-
-public final class UrlUtil {
-    private UrlUtil() {
-    }
-
-    public static String encode(String url, String enc) {
-        try {
-            return URLEncoder.encode(url, enc);
-        } catch (UnsupportedEncodingException e) {
-            throw ExceptionUtil.createException("Could not encode the url: " + url, e);
-        }
-    }
-
-    public static String decode(String url, String enc) {
-        try {
-            return URLDecoder.decode(url, enc);
-        } catch (UnsupportedEncodingException e) {
-            throw ExceptionUtil.createException("Could not decode the url: " + url, e);
-        }
-    }
-}
