--- src/main/java/com/mignonnesaurus/tools/xml/XmlReader.java
@@ -1,4 +1,0 @@
-package com.mignonnesaurus.tools.xml;
-
-public class XmlReader {
-}
