--- src/main/java/com/mignonnesaurus/tools/xml/XmlWriter.java
@@ -1,4 +1,0 @@
-package com.mignonnesaurus.tools.xml;
-
-public class XmlWriter {
-}
