+++ src/test/java/com/github/eliflores/tools/csv/CsvWriterTest.java
@@ -1,0 +1,5 @@
+package com.github.eliflores.tools.csv;
+
+class CsvWriterTest {
+
+}
