+++ src/test/java/com/github/eliflores/tools/xml/XmlReaderTest.java
@@ -1,0 +1,5 @@
+package com.github.eliflores.tools.xml;
+
+class XmlReaderTest {
+
+}
