+++ src/test/java/com/github/eliflores/tools/xml/XmlWriterTest.java
@@ -1,0 +1,5 @@
+package com.github.eliflores.tools.xml;
+
+class XmlWriterTest {
+
+}
