--- src/test/java/com/mignonnesaurus/tools/csv/CsvReaderTest.java
@@ -1,5 +1,0 @@
-package com.mignonnesaurus.tools.csv;
-
-class CsvReaderTest {
-
-}
