--- src/test/java/com/mignonnesaurus/tools/csv/CsvWriterTest.java
@@ -1,5 +1,0 @@
-package com.mignonnesaurus.tools.csv;
-
-class CsvWriterTest {
-
-}
