--- src/test/java/com/mignonnesaurus/tools/xml/XmlWriterTest.java
@@ -1,5 +1,0 @@
-package com.mignonnesaurus.tools.xml;
-
-class XmlWriterTest {
-
-}
