+++ src/test/resources/com/github/eliflores/tools/json/people.json
